feat: dismiss button along with the action button
Always shows dismiss icon doesn't matter action button is available or not.

BREAKING CHANGE: Previously, when action button spec was specified, DISMISS icon-button wasn't ever shown. But, that's not true anymore. If you want to hide action button in such case, you can. There is a property for it, but you need to explicitly set that property.